feat: split backend paths in separate files, simplify backend
